requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
chardet==3.0.4
colorama==0.4.6
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
discord==2.3.2
discord.py==2.4.0
executing==2.1.0
frozenlist==1.4.1
googletrans==3.0.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==1.0.6
httpx==0.27.2
hyperframe==5.2.0
idna==2.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytz==2024.2
pywin32==308
pyzmq==26.2.0
rfc3986==1.5.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
yarl==1.15.4
